requirements.txt
python-dotenv>=1.0.0
httpx>=0.27.0
pandas>=2.0.0
pyodbc>=5.0.0
playwright>=1.40.0
nba_api>=1.4
soccerdata>=1.8
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
